[🥁 : feat] esc키로 닫기 함수 유틸로 분리 후 공통 적용 (#18)
