+++ src/utils/use-esc-close.tsx
@@ -1,0 +1,17 @@
+import { useEffect } from 'react';
+
+const useEscClose = (onClose: () => void) => {
+  useEffect(() => {
+    const handleEscKey = (e: KeyboardEvent) => {
+      if (e.key === 'Escape') {
+        onClose();
+      }
+    };
+    document.addEventListener('keydown', handleEscKey);
+    return () => {
+      document.removeEventListener('keydown', handleEscKey);
+    };
+  }, [onClose]);
+};
+
+export default useEscClose;
